Refactor GolferEndpointTests
