+++ tests/GolfLeague.Api.Tests.Integration/Endpoints/Infrastructure/InfrastructureTests.cs
@@ -11,7 +11,7 @@ public class DatabaseNotNeededFactory : WebApplicationFactory<IGolfApiMarker>
 {
     protected override void ConfigureWebHost(IWebHostBuilder builder)
     {
-        builder.UseSetting("Database:ConnectionString", "Server=localhost,9433;TrustServerCertificate=True;");
+        builder.UseSetting("Database:ConnectionString", "Server=localhost,9433;Connect Timeout=1;TrustServerCertificate=True;");
         base.ConfigureWebHost(builder);
     }
 }
